Some stuff

+++ lib/views/notePage.dart
@@ -1,15 +1,44 @@
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:makemynotes/provider/provider.dart';
 import 'package:makemynotes/views/addPage.dart';
 import 'package:provider/provider.dart';
-import './showPage.dart';
+//import 'showPageOne.dart';
+import 'showPage.dart';
 
 class NotePage extends StatefulWidget {
+  static const routeName = '/notePage';
+
   @override
   _NotePageState createState() => _NotePageState();
 }
 
 class _NotePageState extends State<NotePage> {
+  // deleteNote(String item, String itemTitle) {
+  //   return AlertDialog(
+  //     //title: ,
+  //     content: Text('Delete $itemTitle'),
+  //     actions: [
+  //       TextButton(
+  //         onPressed: () {
+  //           Navigator.of(context).pop();
+  //         },
+  //         child: Text('Cancel'),
+  //       ),
+  //       TextButton(
+  //         onPressed: () {
+  //           Provider.of<TakeNote>(context, listen: false).deleteNote(item);
+  //           print('snak');
+  //           ScaffoldMessenger.of(context)
+  //               .showSnackBar(SnackBar(content: Text('$itemTitle deleted')));
+  //           print('bar');
+  //         },
+  //         child: Text('Delete'),
+  //       ),
+  //     ],
+  //   );
+  // }
+
   @override
   Widget build(BuildContext context) {
     final note = Provider.of<TakeNote>(context, listen: false);
@@ -20,6 +49,7 @@ class _NotePageState extends State<NotePage> {
           IconButton(
               icon: Icon(Icons.add),
               onPressed: () {
+                //Navigator.of(context).pushNamed(AddPage.routeName);
                 Navigator.push(context,
                     MaterialPageRoute(builder: (context) => AddPage()));
               })
@@ -28,6 +58,7 @@ class _NotePageState extends State<NotePage> {
       floatingActionButton: FloatingActionButton(
         child: Icon(Icons.add),
         onPressed: () {
+          //Navigator.of(context).pushNamed(AddPage.routeName);
           Navigator.push(
               context, MaterialPageRoute(builder: (context) => AddPage()));
         },
@@ -50,14 +81,75 @@ class _NotePageState extends State<NotePage> {
                           final itemTitle = takeNote.items[i].title;
                           return Dismissible(
                             key: Key(item),
+                            dragStartBehavior: DragStartBehavior.down,
                             onDismissed: (direction) {
-                              Provider.of<TakeNote>(context, listen: false)
-                                  .deleteNote(item);
-                              print('snak');
+                              print('fvfb');
+                              AlertDialog(
+                                //title: ,
+                                content: Text('Delete $itemTitle'),
+                                actions: [
+                                  TextButton(
+                                    onPressed: () {
+                                      Navigator.of(context).pop();
+                                    },
+                                    child: Text('Cancel'),
+                                  ),
+                                  TextButton(
+                                    onPressed: () {
+                                      Provider.of<TakeNote>(context,
+                                              listen: false)
+                                          .deleteNote(item);
+                                      print('snak');
+                                      ScaffoldMessenger.of(context)
+                                          .showSnackBar(SnackBar(
+                                              content:
+                                                  Text('$itemTitle deleted')));
+                                      print('bar');
+                                    },
+                                    child: Text('Delete'),
+                                  ),
+                                ],
+                              );
+                              print('sfdwfsavcfsqsdac');
                               ScaffoldMessenger.of(context).showSnackBar(
-                                  SnackBar(
-                                      content: Text('$itemTitle deleted')));
-                              print('bar');
+                                SnackBar(
+                                  content: Container(
+                                    child: Column(
+                                      children: <Widget>[
+                                        Text('Do you want to delete this note'),
+                                        Container(
+                                          child: Row(
+                                            children: <Widget>[
+                                              TextButton(
+                                                onPressed: () {
+                                                  Navigator.of(context).pop();
+                                                },
+                                                child: Text('Cancel'),
+                                              ),
+                                              TextButton(
+                                                onPressed: () {
+                                                  Provider.of<TakeNote>(context,
+                                                          listen: false)
+                                                      .deleteNote(item);
+                                                  print('snak');
+                                                  ScaffoldMessenger.of(context)
+                                                      .showSnackBar(SnackBar(
+                                                          content: Text(
+                                                              '$itemTitle deleted')));
+                                                  print('bar');
+                                                },
+                                                child: Text('Delete'),
+                                              ),
+                                            ],
+                                          ),
+                                        ),
+                                      ],
+                                    ),
+                                  ),
+                                ),
+                              );
+                              //deleteNote(item, itemTitle);
+                              print('gnggn');
                             },
                             background: Container(
                               width: MediaQuery.of(context).size.width - 10,
@@ -92,9 +184,17 @@ class _NotePageState extends State<NotePage> {
                                 title: Text(takeNote.items[i].title),
                                 subtitle: takeNote.items[i].text.length >= 20
                                     ? Text(
-                                        '${takeNote.items[i].text.substring(0, 20)}...')
-                                    : Text('${takeNote.items[i].text}'),
+                                        '${takeNote.items[i].text.substring(0, 20)}...',
+                                        maxLines: 1,
+                                      )
+                                    : Text(
+                                        '${takeNote.items[i].text}',
+                                        maxLines: 1,
+                                      ),
                                 onTap: () {
+                                  // Navigator.of(context).pushNamed(
+                                  //     ShowPage.routeName,
+                                  //     arguments: takeNote.items[i].id);
                                   Navigator.push(
                                       context,
                                       MaterialPageRoute(
